capture_mode: Fix a crash in the recording service
In certain failure events, we immediately end recording, and
reset the recording service mojo remote endpoint. This will
end up terminating the recording service process.

- The ~RecordingService() dtor will be called first on the main
  thread.
- Then later the ~RecordingEncoderMuxer() dtor will be called on
  the blocking sequence. This may trigger flushing of any
  non-empty buffers in the muxer. This will trigger callbacks
  on the RecordingService objecty after it had already been
  destroyed.

The solution was to use WeakPtrs to bind these callbacks, but
it's not so simple, since RecordingEncoderMuxer would invoke
them on the blocking sequence, and WeakPtrs cannot be invalidated
except on the sequence they're invoked on.

This CL fixes these issues by:
- Making all callbacks given to RecordingEncoderMuxer bind to
  WeakPtrs and will be invoked on the main thread.
- Handle the case if the client is disconnected by terminating the
  process immediately, since there's no point in doing any extra
  work when no client is out there to receive it.
- Handle the case the RecordingService gets terminated while
  recording is in progress and the client is still connected, by
  stopping the recording, and sending any available cached muxer
  output chunks.

BUG=1160370
TEST=Manually, added browser tests that fail without the fix.

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2801144
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#871625}

+++ chrome/browser/ui/ash/recording_service_browsertest.cc
@@ -151,11 +151,17 @@ class RecordingServiceBrowserTest : public InProcessBrowserTest {
 
   // Reads the video file at the given |path| and verifies its WebM contents. At
   // the end it deletes the file to save space, since video files can be big.
-  void VerifyVideoFileAndDelete(const base::FilePath& path) const {
+  // |allow_empty| can be set to true if an empty video file is possible.
+  void VerifyVideoFileAndDelete(const base::FilePath& path,
+                                bool allow_empty = false) const {
     base::ScopedAllowBlockingForTesting allow_blocking;
     ASSERT_TRUE(base::PathExists(path));
     std::string file_content;
     EXPECT_TRUE(base::ReadFileToString(path, &file_content));
+
+    if (allow_empty && file_content.empty())
+      return;
+
     EXPECT_FALSE(file_content.empty());
     EXPECT_TRUE(WebmVerifier().Verify(file_content));
     EXPECT_TRUE(base::DeleteFile(path));
@@ -251,6 +257,34 @@ IN_PROC_BROWSER_TEST_F(RecordingServiceBrowserTest, RecordRegion) {
   // Select a random partial region of the screen.
   test_api.SetUserSelectedRegion(gfx::Rect(10, 20, 100, 50));
   FinishVideoRecordingTest(&test_api);
+}
+
+IN_PROC_BROWSER_TEST_F(RecordingServiceBrowserTest,
+                       RecordingServiceEndpointDropped) {
+  ash::CaptureModeTestApi test_api;
+  test_api.StartForFullscreen(/*for_video=*/true);
+  test_api.PerformCapture();
+  test_api.FlushRecordingServiceForTesting();
+  WaitForMilliseconds(1000);
+  test_api.ResetRecordingServiceRemote();
+  const base::FilePath video_path = WaitForVideoFileToBeSaved();
+  VerifyVideoFileAndDelete(video_path);
+}
+
+IN_PROC_BROWSER_TEST_F(RecordingServiceBrowserTest,
+                       RecordingServiceClientEndpointDropped) {
+  ash::CaptureModeTestApi test_api;
+  test_api.StartForFullscreen(/*for_video=*/true);
+  test_api.PerformCapture();
+  test_api.FlushRecordingServiceForTesting();
+  WaitForMilliseconds(1000);
+  test_api.ResetRecordingServiceClientReceiver();
+  const base::FilePath video_path = WaitForVideoFileToBeSaved();
+  // Due to buffering on the service side, the channel might get dropped before
+  // any flushing of those beffers ever happens, and since dropping the client
+  // end point will immediately terminate the service, nothing may ever get
+  // flushed, and the resulting video file can be empty.
+  VerifyVideoFileAndDelete(video_path, /*allow_empty=*/true);
 }
 
 // Doing multiple recordings one after the other should produce non-corrupt webm
